ui-proxy/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
jinja2>=3.1
requests>=2.32
google-auth>=2.34
